requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
google-generativeai>=0.3.0
python-dotenv>=0.21.0
